+++ src/schema/mutation/addGroup.mutation.ts
@@ -4,6 +4,13 @@ import { AppAbility } from '@security/defineUserAbility';
 import { GroupType } from '../types';
 import config from 'config';
 import { logger } from '@services/logger.service';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the addGroup mutation */
+type AddGroupArgs = {
+  title: string;
+};
 
 /**
  * Creates a new group.
@@ -14,14 +21,10 @@ export default {
   args: {
     title: { type: new GraphQLNonNull(GraphQLString) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: AddGroupArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
       const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
 
       if (!config.get('user.groups.local')) {
         throw new GraphQLError(