+++ src/schema/mutation/deleteApplication.mutation.ts
@@ -6,6 +6,14 @@ import channels from '@const/channels';
 import { AppAbility } from '@security/defineUserAbility';
 import { logger } from '@services/logger.service';
 import { accessibleBy } from '@casl/mongoose';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Types } from 'mongoose';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the deleteApplication mutation */
+type DeleteApplicationArgs = {
+  id: string | Types.ObjectId;
+};
 
 /**
  * Deletes an application from its id.
@@ -17,15 +25,9 @@ export default {
   args: {
     id: { type: new GraphQLNonNull(GraphQLID) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: DeleteApplicationArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
-      // Authentication check
-      const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
       // Delete the application
       const ability: AppAbility = context.user.ability;
       const filters = Application.find(