+++ src/schema/mutation/duplicatePage.mutation.ts
@@ -4,6 +4,16 @@ import { PageType } from '../types';
 import { Application, Page, Role, Step, Workflow } from '@models';
 import { duplicatePage } from '../../services/page.service';
 import { logger } from '@services/logger.service';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Types } from 'mongoose';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the duplicatePage mutation */
+type DuplicatePageArgs = {
+  page?: string | Types.ObjectId;
+  step?: string | Types.ObjectId;
+  application: string | Types.ObjectId;
+};
 
 /**
  * Duplicate existing page in a new application.
@@ -17,15 +27,11 @@ export default {
     step: { type: GraphQLID },
     application: { type: new GraphQLNonNull(GraphQLID) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: DuplicatePageArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
       // Check ability
       const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
       const ability: AppAbility = user.ability;
       // Check parameters
       if (!args.page && !args.step) {