+++ src/schema/mutation/publish.mutation.ts
@@ -10,6 +10,15 @@ import { AppAbility } from '@security/defineUserAbility';
 import pubsubSafe from '../../server/pubsubSafe';
 import config from 'config';
 import { logger } from '@services/logger.service';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Types } from 'mongoose';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the publish mutation */
+type PublishArgs = {
+  ids: string[] | Types.ObjectId[];
+  channel: string | Types.ObjectId;
+};
 
 /**
  * Publish records in a notification.
@@ -20,16 +29,9 @@ export default {
     ids: { type: new GraphQLNonNull(new GraphQLList(GraphQLID)) },
     channel: { type: new GraphQLNonNull(GraphQLID) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: PublishArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
-      // Authentication check
-      const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
-
       const ability: AppAbility = context.user.ability;
       const channel = await Channel.findById(args.channel).populate({
         path: 'application',