+++ src/schema/mutation/seeNotifications.mutation.ts
@@ -9,6 +9,14 @@ import { Notification } from '@models';
 import { AppAbility } from '@security/defineUserAbility';
 import { logger } from '@services/logger.service';
 import { accessibleBy } from '@casl/mongoose';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Types } from 'mongoose';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the seeNotifications mutation */
+type SeeNotificationsArgs = {
+  ids: string[] | Types.ObjectId[];
+};
 
 /**
  * Find multiple notifications and mark them as read.
@@ -19,15 +27,11 @@ export default {
   args: {
     ids: { type: new GraphQLNonNull(new GraphQLList(GraphQLID)) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: SeeNotificationsArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
       // Authentication check
       const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
 
       const ability: AppAbility = context.user.ability;
       if (!args) {