+++ src/schema/query/dashboard.query.ts
@@ -11,6 +11,13 @@ import extendAbilityForContent from '@security/extendAbilityForContent';
 import { CustomAPI } from '@server/apollo/dataSources';
 import { Types } from 'mongoose';
 import { logger } from '@services/logger.service';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Context } from '@server/apollo/context';
+
+/** Arguments for the dashboard query */
+type DashboardArgs = {
+  id: string | Types.ObjectId;
+};
 
 /**
  * Return dashboard from id if available for the logged user.
@@ -21,16 +28,10 @@ export default {
   args: {
     id: { type: new GraphQLNonNull(GraphQLID) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: DashboardArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
-      // Authentication check
       const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
-
       // get data and check permissions
       const dashboard = await Dashboard.findById(args.id);
       const ability = await extendAbilityForContent(user, dashboard);