+++ src/schema/query/group.query.ts
@@ -4,7 +4,14 @@ import { GroupType } from '../types';
 import { AppAbility } from '@security/defineUserAbility';
 import { logger } from '@services/logger.service';
 import { accessibleBy } from '@casl/mongoose';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Types } from 'mongoose';
+import { Context } from '@server/apollo/context';
 
+/** Arguments for the group query */
+type GroupArgs = {
+  id: string | Types.ObjectId;
+};
 /**
  * Get Query by ID.
  * Throw error if user is not logged, or does not have permission to see group, or group does not exist.
@@ -14,16 +21,9 @@ export default {
   args: {
     id: { type: new GraphQLNonNull(GraphQLID) },
   },
-  async resolve(parent, args, context) {
+  async resolve(parent, args: GroupArgs, context: Context) {
+    graphQLAuthCheck(context);
     try {
-      // Authentication check
-      const user = context.user;
-      if (!user) {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
-
       const ability: AppAbility = context.user.ability;
       if (ability.can('read', 'Group')) {
         try {