+++ src/schema/query/me.query.ts
@@ -1,6 +1,8 @@
 import { GraphQLError } from 'graphql';
 import { UserType } from '../types';
 import { logger } from '@services/logger.service';
+import { graphQLAuthCheck } from '@schema/shared';
+import { Context } from '@server/apollo/context';
 
 /**
  * Return user from logged user id.
@@ -8,16 +10,10 @@ import { logger } from '@services/logger.service';
  */
 export default {
   type: UserType,
-  resolve: async (parent, args, context) => {
+  resolve: async (parent, args, context: Context) => {
+    graphQLAuthCheck(context);
     try {
-      const user = context.user;
-      if (user) {
-        return user;
-      } else {
-        throw new GraphQLError(
-          context.i18next.t('common.errors.userNotLogged')
-        );
-      }
+      return context.user;
     } catch (err) {
       logger.error(err.message, { stack: err.stack });
       if (err instanceof GraphQLError) {